feat(routing): adding basic routes
- Main route for overall graph
- Second route skeleton for stats
