+++ package.json
@@ -14,6 +14,7 @@
     "react": "^16.13.1",
     "react-app-rewired": "^2.1.5",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0",
     "recharts": "^1.8.5"
   },
