+++ src/routes/stats/Stats.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Stats = () => {
+    return (
+        <div>
+            secound route for stats!
+        </div>
+    )
+}
+
+export default Stats
